Updated recipe to match the "New Readability"'s page formats
* calibre-readability.recipe NOTES
This works now, but is picking up less metadata than the old recipe,
doesn’t seem to be able to read the Favorites list, and some of the CSS
selectors are rather clunkily done because I’ve just started learning
beautifulsoup.  I figure it’s at least a usable patch, but it does need
more work.

* debug.sh NOTES
Just had to change the last line so it would find Calibre on my
machine. This is presumably machine-specific, but mine is at least the
default Mac OS install path.